# frontend/src/types/openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "人生草稿本 API",
    "description": "Life Draft Book — Interactive narrative game API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/character/setting": {
      "post": {
        "tags": [
          "Character"
        ],
        "summary": "Generate Setting",
        "description": "Generate a character setting (era, age, gender, world, family, relationships, traits).",
        "operationId": "generate_setting_api_character_setting_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSettingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/character/relationship": {
      "post": {
        "tags": [
          "Character"
        ],
        "summary": "Generate Relationship",
        "description": "Generate a single relationship person with rich attributes.",
        "operationId": "generate_relationship_api_character_relationship_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRelationshipRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/character/attributes": {
      "post": {
        "tags": [
          "Character"
        ],
        "summary": "Generate Attributes",
        "description": "Generate initial character attributes (energy, mood, knowledge).",
        "operationId": "generate_attributes_api_character_attributes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateAttributesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/character/opening-story": {
      "post": {
        "tags": [
          "Character"
        ],
        "summary": "Generate Opening Story",
        "description": "Generate opening story via SSE streaming.",
        "operationId": "generate_opening_story_api_character_opening_story_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpeningStoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/character/relationships-summary": {
      "post": {
        "tags": [
          "Character"
        ],
        "summary": "Generate Relationships Summary",
        "description": "Generate a narrative summary of all relationships.",
        "operationId": "generate_relationships_summary_api_character_relationships_summary_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationshipsSummaryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/collection/{game_id}/characters/{character_name}": {
      "delete": {
        "tags": [
          "Collection"
        ],
        "summary": "Delete Character",
        "description": "删除人物",
        "operationId": "delete_character_api_collection__game_id__characters__character_name__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Game Id"
            }
          },
          {
            "name": "character_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenerateAttributesRequest": {
        "properties": {
          "character_settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Character Settings",
            "x-maxBytes": 262144
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "zh"
          }
        },
        "type": "object",
        "required": [
          "character_settings"
        ],
        "title": "GenerateAttributesRequest"
      },
...
      "GenerateRelationshipRequest": {
        "properties": {
          "player_name": {
            "type": "string",
            "maxLength": 50,
            "title": "Player Name"
          },
          "life_vision": {
            "type": "string",
            "maxLength": 500,
            "title": "Life Vision"
          },
          "previous_settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Previous Settings",
            "x-maxBytes": 262144
          },
          "existing_people": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Existing People"
          },
          "person_index": {
            "type": "integer",
            "title": "Person Index",
            "default": 0
          },
          "total_needed": {
            "type": "integer",
            "title": "Total Needed",
            "default": 3
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "zh"
          }
        },
        "type": "object",
        "required": [
          "player_name",
          "life_vision"
        ],
        "title": "GenerateRelationshipRequest"
      },
...
      "GenerateSettingRequest": {
        "properties": {
          "setting_type": {
            "type": "string",
            "enum": [
              "era",
              "age",
              "gender",
              "world",
              "family",
              "relationships",
              "traits"
            ],
            "title": "Setting Type",
            "description": "era|age|gender|world|family|relationships|traits"
          },
          "player_name": {
            "type": "string",
            "maxLength": 50,
            "title": "Player Name"
          },
          "life_vision": {
            "type": "string",
            "maxLength": 500,
            "title": "Life Vision"
          },
          "previous_settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Previous Settings",
            "x-maxBytes": 262144
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "zh"
          }
        },
        "type": "object",
        "required": [
          "setting_type",
          "player_name",
          "life_vision"
        ],
        "title": "GenerateSettingRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OpeningStoryRequest": {
        "properties": {
          "character_settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Character Settings",
            "x-maxBytes": 262144
          },
          "player_name": {
            "type": "string",
            "maxLength": 50,
            "title": "Player Name"
          },
          "life_vision": {
            "type": "string",
            "maxLength": 500,
            "title": "Life Vision"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "zh"
          }
        },
        "type": "object",
        "required": [
          "character_settings",
          "player_name",
          "life_vision"
        ],
        "title": "OpeningStoryRequest"
      },
...
      "RelationshipsSummaryRequest": {
        "properties": {
          "player_name": {
            "type": "string",
            "maxLength": 50,
            "title": "Player Name"
          },
          "life_vision": {
            "type": "string",
            "maxLength": 500,
            "title": "Life Vision"
          },
          "previous_settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Previous Settings",
            "x-maxBytes": 262144
          },
          "key_people": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Key People"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "zh"
          }
        },
        "type": "object",
        "required": [
          "player_name",
          "life_vision"
        ],
        "title": "RelationshipsSummaryRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}